contrib/IBM/sarama: add WrapConsumerGroupHandler to trace consumer groups
